refactor: movieInfoTitle 분리
