+++ BoxOffice/Model/MovieInfoTitle.swift
@@ -1,0 +1,19 @@
+//
+//  MovieInfoTitle.swift
+//  BoxOffice
+//
+//  Created by Seoyeon Hong on 2023/04/05.
+//
+
+import Foundation
+
+enum MovieInfoTitle {
+    static let director = "감독"
+    static let productionYear = "제작년도"
+    static let openDate = "개봉일"
+    static let showTime = "상영시간"
+    static let watchGradeName = "관람등급"
+    static let nationName = "제작국가"
+    static let genreName = "장르"
+    static let actorName = "배우"
+}
