+++ BoxOffice/View/DetailMovieInfoViewController.swift
@@ -9,17 +9,6 @@ import UIKit
 
 final class DetailMovieInfoViewController: UIViewController {
     
-    enum MovieInfoTitle {
-        static let director = "감독"
-        static let productionYear = "제작년도"
-        static let openDate = "개봉일"
-        static let showTime = "상영시간"
-        static let watchGradeName = "관람등급"
-        static let nationName = "제작국가"
-        static let genreName = "장르"
-        static let actorName = "배우"
-    }
-    
     private let movieCode: String
     private let movieName: String
     private let provider = APIProvider.shared
